pyproject.toml
[tool.poetry]
name = "pingis-ai"
version = "0.1.0"
description = "A Retrieval-Augmented Generation (RAG) pipeline integrating a vector database for efficient semantic search and context-aware responses."
authors = ["Daniel Haglund <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
python-dotenv = "^1.0.0"
langchain-core = ">=0.3.34,<1.0.0"
langchain-text-splitters = "^0.3.6"
langchain = "^0.3.18"
langgraph = "^0.2.70"
langsmith = "^0.3.8"
qdrant-client = "^1.0.0"
langchain-qdrant = "^0.1.0"
langchain-openai = "^0.3.5"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
graphviz = "^0.20.1"
aiofiles = "^24.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"